Recentered all the art prefabs, touchups in colliders & pressure plates in rooms, + part 1 of temporary animations
Altered files:
- MicrowavePrototyping01 scene
- Almost every prefab in microwave art folder, excluding the room prefabs
- created animations for gates and fans
