+++ Assets/LevelObjects/Art/Prefabs/[TempSmallScrew].prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a53095fb154cfa4fb461e207a5335cf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
