+++ Assets/LevelObjects/Scripts/[BraydenTestScript].meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba19264170019234e8110a844e41678c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
