+++ Assets/LevelObjects/Scripts/[BraydenTestScript]/TestIndicators.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 057a79fccc815ac4a95f367e964d9ec8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
