+++ Assets/Levels/Microwave/Art/Prefabs/Fans.prefab
@@ -62,17 +62,17 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: 3a0bfd4c927bef04aa477338edcd60de,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: -6.33334
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3a0bfd4c927bef04aa477338edcd60de,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -1.56
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3a0bfd4c927bef04aa477338edcd60de,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: -21.66
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3a0bfd4c927bef04aa477338edcd60de,
         type: 3}
