+++ Assets/Levels/Microwave/Art/Prefabs/GateL.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 77693034e870a884a80b11af60753cab
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
