+++ Assets/Levels/Microwave/Art/Prefabs/GateR.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e382ea13ef193a44e89d25d868af30e3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
