+++ Assets/Levels/Microwave/Art/Prefabs/HeatFence.prefab
@@ -23,15 +23,15 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3785067743064578270}
-  m_LocalRotation: {x: -0, y: -0.35490856, z: -0, w: 0.93490106}
-  m_LocalPosition: {x: 5.89, y: -3.94, z: -9.48}
-  m_LocalScale: {x: 0.94930935, y: 1, z: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1572261536761844571}
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: -41.576, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1322777311720666288
 PrefabInstance:
   m_ObjectHideFlags: 0
