+++ Assets/Levels/Microwave/Art/Prefabs/Mesh_Cube.prefab
@@ -47,17 +47,17 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: 781ccf35fffc5244b9a719a0d31e42b2,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: 3.96
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 781ccf35fffc5244b9a719a0d31e42b2,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -0.68
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 781ccf35fffc5244b9a719a0d31e42b2,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: -18.12
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 781ccf35fffc5244b9a719a0d31e42b2,
         type: 3}
