+++ Assets/Levels/Microwave/Art/Prefabs/PowerTower (1).prefab
@@ -23,15 +23,15 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1018590782428553692}
-  m_LocalRotation: {x: -0, y: -0.18592645, z: -0, w: 0.9825637}
-  m_LocalPosition: {x: -6.01, y: -2.66, z: -14.27}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1336263570701137176}
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: -21.43, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1513743692481090291
 PrefabInstance:
   m_ObjectHideFlags: 0
