+++ Assets/Levels/Microwave/Art/Prefabs/PowerTower.prefab
@@ -23,15 +23,15 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6358586684299266276}
-  m_LocalRotation: {x: -0, y: -0.18592645, z: -0, w: 0.9825637}
-  m_LocalPosition: {x: -6.14, y: -2.66, z: -9.96}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 8745348566955830182}
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: -21.43, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &9139775423371201613
 PrefabInstance:
   m_ObjectHideFlags: 0
