+++ Assets/Levels/Microwave/Art/Prefabs/ScrewHole.prefab
@@ -28,14 +28,14 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2104686175981012972}
-  m_LocalRotation: {x: 0.49289182, y: -0.48942208, z: 0.5068865, w: 0.51048005}
-  m_LocalPosition: {x: -0.012018204, y: 0.62802505, z: 0.581995}
+  m_LocalRotation: {x: 0.5, y: -0.5, z: 0.5, w: 0.5}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 0.14730066, y: 0.041568264, z: 0.1473007}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 8457730952069389080}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 87.992, y: 0, z: 89.595}
+  m_LocalEulerAnglesHint: {x: 90, y: 0, z: 90}
 --- !u!114 &2104686175981012970
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2231,7 +2231,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8097215154729512503}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -20.215982, y: 18.761974, z: 25.159004}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
