+++ Assets/Levels/Microwave/Art/Prefabs/SteelBeam.prefab
@@ -24,7 +24,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3265849328431363951}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 48.84873, y: 23.056845, z: -41.334282}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -47,17 +47,17 @@ PrefabInstance:
     - target: {fileID: -8679921383154817045, guid: 3c9aa2f0c9951404685c9644ff33b687,
         type: 3}
       propertyPath: m_LocalPosition.x
-      value: -48.84873
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3c9aa2f0c9951404685c9644ff33b687,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: -23.056845
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3c9aa2f0c9951404685c9644ff33b687,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: 41.334282
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: -8679921383154817045, guid: 3c9aa2f0c9951404685c9644ff33b687,
         type: 3}
