+++ Assets/Levels/Microwave/LevelObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c93d77b9b2c0ee4dba8870d16eed52a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
