+++ Assets/Levels/Microwave/LevelObjects/Fan.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f26374662c2ad648bd5cfef59ec17e9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
