+++ Assets/Levels/Microwave/LevelObjects/Fan.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eab40af44654a324e8e0ba1c7c2a6175
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
