+++ Assets/Levels/Microwave/PrototypingAnimations/Gates.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7760798cee8c1a5499efbc39d26f3c9a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
