+++ Assets/Levels/Microwave/PrototypingAnimations/OnActivate.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6cb1a8a6a8b52e41a186174a7a693de
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
